Refactor bridging_packets tests
